Add a simple test for multiple
